refactor: split large components into modular structure and add unit tests
- Refactor useModelOperations into modular files
- Refactor CommonProTable, ModelCustom, ModelList components
- Refactor ModelDetail and useInlineOperations into sub-modules
- Add comprehensive unit tests for ModelDetail modules
- Fix lint warnings and update test snapshots

No changes to original API or business logic
