+++ src/components/ModelDetail/useInlineOperations/types.ts
@@ -1,0 +1,11 @@
+export interface UseInlineOperationsOptions {
+  mainRecord: Record<string, any>;
+}
+
+export interface InlineDataState {
+  [key: string]: Record<string, any>[];
+}
+
+export interface EditingKeysState {
+  [key: string]: any[];
+}
