+++ src/components/ModelDetail/utils.ts
@@ -1,0 +1,4 @@
+// Utility function: capitalize first letter
+export const capitalizeFirstLetter = (str: string) => {
+  return str.charAt(0).toUpperCase() + str.slice(1);
+};
